// apps/destiny-clan-dashboard/src/app/features/recent-activity/models/profile-recent-activity.ts
import { ActivityStats, MemberProfile } from '@destiny-clan-dashboard/data/models';
import { MemberActivityTime } from 'projects/data/src/lib/models/MemberActivityTime';
import { GroupsV2GroupMember } from 'bungie-api-angular';

export interface ProfileRecentActivity {
  clan: {
    clanId: string;
    clanName: string;
    clanTag: string;
  };
  clanMember: GroupsV2GroupMember;
  profile: MemberProfile;
  profileActivity: ActivityStats;
}